refactor: reorganize utility imports and centralize validation rules in shared utils
