+++ src/shared/utils/index.ts
@@ -14,17 +14,23 @@ export { convertTransparencyToHex, sortFields, splitName } from './stringUtils';
 
 export {
 	confirmPasswordRule,
-	getPasswordChecks,
+	// getPasswordChecks,
 	hasSpecialCharRule,
 	minLengthRule,
 	passwordValidationRule,
 	requiredFieldRule,
-	validateEmails,
+	// validateEmails,
 	validateEmailsRule,
 	validEmailRule,
 } from './validators';
 export type { ValidationRule } from './validators';
 
-export { buildDocumentLinkUrl, buildResetPasswordUrl, buildVerificationUrl } from './urlBuilder';
+export {
+	buildDocumentLinkUrl,
+	buildResetPasswordUrl,
+	buildVerificationUrl,
+} from './urlBuilderUtils';
 
 export { emptyToNull, parseBoolean, dedupeByEmail, handleEmailSelection } from './formDataUtils';
+
+export { PASSWORD_RULES, getPasswordChecks, validateEmails, isValidEmail } from './validationUtils';
