+++ src/shared/utils/validationUtils.ts
